chore(Section): extract generateMeta function into dedicated file
